# specs/002-http-download-api/contracts/openapi.yaml
openapi: 3.1.0

info:
  title: X Video Downloader API
  version: 1.0.0
  description: |
    Submit an X post URL, receive a job handle immediately, poll for state, and collect the file.

    **Authorization is the handle.** Possession of a job's handle is what entitles a caller to that
    job; there is no other credential (spec Q1, FR-028). Handles carry 256 bits of entropy.

    **No error response carries a filesystem path, a directory name, a stack trace, or the text of
    an underlying library's error** (FR-029). Every message in this contract is a fixed literal.

servers:
  - url: http://localhost:8000
    description: Local. TLS, domain, and reverse proxy are deployment concerns, out of scope.

paths:
  /jobs:
    post:
      summary: Submit a post URL
      description: |
        Returns without waiting for the download (FR-001). The URL is validated by the frozen
        `validation.parse_post_url` before a job exists; an invalid URL produces no job and no
        outbound network request (FR-002, FR-003).

        A submission naming a post that is already `waiting` or `running` returns that job's
        existing handle rather than starting a second download (FR-016).
      operationId: submitJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitRequest'
      responses:
        '202':
          description: Accepted. Body is identical for a new job and a deduplicated one.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: |
            The URL failed validation. The message is a fixed literal and never echoes the
            submitted value (FR-005).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: invalid_url, message: "That is not a valid X post URL." }
        '422':
          description: |
            Malformed body. The default pydantic handler is replaced so the offending input is
            never reflected back (FR-005, research D6).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: invalid_request, message: "The request body was not understood." }
        '429':
          description: |
            Rate limit exceeded for the calling address (FR-019).

            Counted per *submission*, not per job created: an invalid URL and a deduplicated
            submission both consume allowance. The address comes from the connection and never
            from a header — see research.md D9 for which proxy configurations change that.
          headers:
            Retry-After:
              schema: { type: integer }
              description: |
                Seconds until the caller may submit again. Computed from the service's own
                clock and window, so it discloses nothing about the request or the server.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: rate_limited
                message: "Too many submissions. Try again in 240 seconds."
        '503':
          description: |
            Two distinct refusals share this status, both meaning "not now, and no job was
            created":

            * `insufficient_storage` — free disk on the output volume is below the configured
              threshold (FR-018);
            * `at_capacity` — the pending queue is at `XVD_MAX_PENDING` (FR-015 as amended).

            Neither message names the volume, the threshold, the queue depth, or how many jobs
            are pending. Those are facts about the server (FR-029).

            A deduplicated submission is never refused by either: it creates no job and consumes
            no disk, so the caller receives the existing job's handle even at capacity.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                insufficientStorage:
                  value:
                    code: insufficient_storage
                    message: "The service is temporarily out of space. Try again later."
                atCapacity:
                  value:
                    code: at_capacity
                    message: "The service is at capacity. Try again shortly."

  /jobs/{handle}:
    get:
      summary: Read a job's state
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/Handle'
      responses:
        '200':
          description: Current state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404':
          $ref: '#/components/responses/NotFound'

  /jobs/{handle}/file:
    get:
      summary: Retrieve the file of a single-file job
      description: |
        Returns the file when the job produced exactly one — the common case, which requires no
        index (FR-035, spec Q2). When the job produced several, this refuses with `409` naming the
        count, and the caller retries against `/jobs/{handle}/file/{index}`.
      operationId: getJobFile
      parameters:
        - $ref: '#/components/parameters/Handle'
      responses:
        '200':
          description: |
            The file exactly as downloaded — no transcoding (FR-013). `Content-Disposition` carries
            a server-derived filename; the caller never influences it.
          content:
            video/mp4:
              schema: { type: string, format: binary }
            application/octet-stream:
              schema: { type: string, format: binary }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            The job is not in a state that yields this file. Which state applies is named, because
            holding the handle already entitles the caller to know it (FR-014).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                notReady:
                  value: { code: not_ready, message: "This job has not finished yet." }
                failed:
                  value: { code: failed, message: "This post does not contain a video." }
                multipleFiles:
                  value:
                    code: index_required
                    message: "This job produced 3 files. Request one by index."
        '410':
          $ref: '#/components/responses/Expired'

  /jobs/{handle}/file/{index}:
    get:
      summary: Retrieve one file of a multi-file job
      operationId: getJobFileByIndex
      parameters:
        - $ref: '#/components/parameters/Handle'
        - name: index
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
          description: 1-based, matching the numbering the frozen downloader uses in filenames.
      responses:
        '200':
          description: The file exactly as downloaded.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404':
          description: |
            Unknown handle, **or** an index outside the job's range (FR-036). Deliberately the same
            response — the count is already available from `GET /jobs/{handle}`, so distinguishing
            them here would add nothing.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: The job has not finished.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '410':
          $ref: '#/components/responses/Expired'

  /health:
    get:
      summary: Operator liveness and capacity
      description: |
        Carries no caller data and no job handles. `wedged_workers` surfaces the known limitation in
        research D4 — a download whose `ffmpeg` merge hung holds a worker thread until restart.
      operationId: health
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [status, running, waiting, wedged_workers]
                properties:
                  status: { type: string, enum: [ok, degraded] }
                  running: { type: integer }
                  waiting: { type: integer }
                  wedged_workers:
                    type: integer
                    description: Workers whose job the watchdog failed but which never returned.

components:
  parameters:
    Handle:
      name: handle
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{43}$'
      description: |
        The capability. Never used to build a filesystem path — lookup is an in-memory dict hit, so
        a caller-supplied value cannot reach the filesystem at all (research D3).

  responses:
    NotFound:
      description: |
        One fixed response for every handle that does not resolve — unknown, malformed, or already
        purged. A caller cannot learn whether a handle names a real job (FR-028).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: not_found, message: "No such job." }
    Expired:
      description: The file was deleted after the retention period (FR-021, FR-022).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: expired
            message: "This file has expired and is no longer available."

  schemas:
    SubmitRequest:
      type: object
      required: [url]
      additionalProperties: false
      description: |
        `additionalProperties: false` is load-bearing, not tidiness. It is what makes FR-004
        enforceable at the boundary: no field naming an output location, a target file, or a format
        can be smuggled in and silently honoured.
      properties:
        url:
          type: string
          maxLength: 2048
          description: A single X post URL. Length-capped so an oversized body is refused (FR-005).

    Job:
      type: object
      required: [handle, state, file_count]
      description: |
        Everything a caller may see. Notably absent: any filesystem path, the output directory, the
        raw text of a download failure, and the calling address. The absence is structural — the
        service-layer record holds no such field (research D6).
      properties:
        handle:
          type: string
          example: FQPTq5wsoSUHUoCYabEc8Ptkfl3tyVZaHYG_J-n-XRY
        state:
          type: string
          enum: [waiting, running, finished, failed, expired]
        file_count:
          type: integer
          description: |
            0 until finished. When >1, retrieval requires an index (FR-012, FR-035).
        progress:
          $ref: '#/components/schemas/Progress'
        failure:
          $ref: '#/components/schemas/Failure'
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }

    Progress:
      type: object
      nullable: true
      description: |
        Present only while running, and **advisory** (FR-008). `total_bytes` is often absent because
        yt-dlp does not always report one, and for a multi-video post the figures restart per video.
        A caller must not treat this as monotonic or as a completion guarantee.
      properties:
        downloaded_bytes: { type: integer }
        total_bytes: { type: integer, nullable: true }
        percent: { type: number, nullable: true }

    Failure:
      type: object
      nullable: true
      description: Present only when `state` is `failed`.
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - no_video
            - not_a_video
            - protected_account
            - age_restricted
            - post_unavailable
            - interrupted
            - service_unavailable
            - time_limit
            - unclassified
          description: |
            Stable and machine-readable (FR-009, FR-010). `unclassified` is a deliberate, visible
            outcome, never a silent substitution (FR-011).
        message:
          type: string
          description: A fixed sentence from the server's own catalog. Never derived from any
            library's error text (FR-029).

    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code: { type: string }
        message:
          type: string
          description: A fixed literal. Never contains a path, a directory name, a stack trace, or
            the caller's submitted content.